letsencrypt: add graphite.opendev.org challenge CNAME
This is for initial testing of letsencrypt; it creates CNAMES for the
the ACME challenge entries; the TXT records will be created in
acme.opendev.org by Ansible to allow for certificate generation.

Addtionally we remove the IN NS for acme.opendev.org; as pointed out
in prior reviews it is unnecessary so done to avoid confusion.

Change-Id: Ibe48384a4a09f858c36ac0a305656444181f21d6