Add RFC 6844 CAA RR for graphite01
The DNS Certification Authority Authorization (CAA) Resource Record
described in IETF RFC 6844 allows us to specify which certificate
authorities we expect to issue certificates for a given hostname.
This is a measure to indicate to all reputable CAs that they should
not honor any request for a certificate unless they are one of the
parties listed. In this case, assert that only letsencrypt.org is
expected to issue certificates for the graphite CNAME, along with an
E-mail address to which any identified policy violations should be
reported.

Change-Id: I7ccb3a177386085221f0c85b370c08fcf031703e